hub/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.3
webauthn==2.2.0
slowapi==0.1.9
